rename to generic
