+++ pycaret/anomaly/functional.py
@@ -7,7 +7,7 @@ from joblib.memory import Memory
 from pycaret.anomaly.oop import AnomalyExperiment
 from pycaret.loggers.base_logger import BaseLogger
 from pycaret.utils.constants import DATAFRAME_LIKE, SEQUENCE_LIKE
-from pycaret.utils.utils import check_if_global_is_not_none
+from pycaret.utils.generic import check_if_global_is_not_none
 
 _EXPERIMENT_CLASS = AnomalyExperiment
 _CURRENT_EXPERIMENT: Optional[AnomalyExperiment] = None
