+++ pycaret/containers/models/regression.py
@@ -25,7 +25,7 @@ from pycaret.internal.distributions import (
     UniformDistribution,
 )
 from pycaret.utils._dependencies import _check_soft_dependencies
-from pycaret.utils.utils import get_logger, np_list_arange, param_grid_to_lists
+from pycaret.utils.generic import get_logger, np_list_arange, param_grid_to_lists
 
 # First one in the list is the default ----
 ALL_ALLOWED_ENGINES: Dict[str, List[str]] = {
