+++ pycaret/internal/display/__init__.py
@@ -1,4 +1,4 @@
 from pycaret.internal.display.display import CommonDisplay, DummyDisplay
-from pycaret.utils.utils import enable_colab
+from pycaret.utils.generic import enable_colab
 
 __all__ = ["CommonDisplay", "DummyDisplay", "enable_colab"]
