+++ pycaret/internal/pipeline.py
@@ -24,7 +24,7 @@ from sklearn.utils.metaestimators import if_delegate_has_method
 from sklearn.utils.validation import check_memory
 
 from pycaret.utils._show_versions import _get_deps_info
-from pycaret.utils.utils import get_all_object_vars_and_properties, variable_return
+from pycaret.utils.generic import get_all_object_vars_and_properties, variable_return
 
 
 def _fit_one(transformer, X=None, y=None, message=None, **fit_params):
