+++ pycaret/internal/plots/time_series.py
@@ -23,8 +23,8 @@ from pycaret.internal.plots.utils.time_series import (
     qq_subplot,
     time_series_subplot,
 )
+from pycaret.utils.generic import _resolve_dict_keys
 from pycaret.utils.time_series import get_diffs
-from pycaret.utils.utils import _resolve_dict_keys
 
 __author__ = ["satya-pattnaik", "ngupta23"]
