+++ pycaret/time_series/forecasting/oop.py
@@ -62,6 +62,12 @@ from pycaret.internal.validation import is_sklearn_cv_generator
 from pycaret.loggers.base_logger import BaseLogger
 from pycaret.utils._dependencies import _check_soft_dependencies
 from pycaret.utils.datetime import coerce_datetime_to_period_index
+from pycaret.utils.generic import (
+    MLUsecase,
+    _coerce_empty_dataframe_to_none,
+    _resolve_dict_keys,
+    highlight_setup,
+)
 from pycaret.utils.time_series import (
     TSApproachTypes,
     TSExogenousPresent,
@@ -84,12 +90,6 @@ from pycaret.utils.time_series.forecasting.pipeline import (
     _add_model_to_pipeline,
     _get_imputed_data,
     _get_pipeline_estimator_label,
-)
-from pycaret.utils.utils import (
-    MLUsecase,
-    _coerce_empty_dataframe_to_none,
-    _resolve_dict_keys,
-    highlight_setup,
 )
 
 LOGGER = get_logger()
