+++ tests/test_regression_tuning.py
@@ -8,7 +8,7 @@ import pytest
 
 import pycaret.datasets
 import pycaret.regression
-from pycaret.utils.utils import can_early_stop
+from pycaret.utils.generic import can_early_stop
 
 
 @pytest.mark.skip(reason="no way of currently testing this")
